More EXTRACT_8BITS() and other cleanups.
Get rid of casts to (int) that aren't needed or wanted.

If a field is unsigned, use an unsigned variable for it, print it with
%u, not %d, and don't cast it to int.

Replace a static variable in print-dvmrp.c with a local variable in
dvmrp_print() and a parameter to print_neighbors2().

+++ print-dhcp6.c
@@ -665,7 +665,7 @@ dhcp6opt_print(netdissect_options *ndo,
 				ND_PRINT((ndo, " by-clientID"));
 				break;
 			default:
-				ND_PRINT((ndo, " type_%d", (int)dh6_lq_query_type));
+				ND_PRINT((ndo, " type_%u", dh6_lq_query_type));
 				break;
 			}
 			ND_PRINT((ndo, " %s", ip6addr_string(ndo, &tp[1])));
